requirements.txt
bittensor==6.9.4
datasets==2.*
envparse==0.2.0
openai>=1.3.2, ==1.*
Pillow==10.*
requests==2.*
scikit-learn==1.*
torch==2.*
transformers==4.*
wandb
anthropic==0.19.2
stability-sdk
boto3
anthropic_bedrock
pyOpenSSL==24.*
google-generativeai
groq==0.5.0
aioboto3